requirements.txt
Pillow>=8.4.0
ffmpeg-python>=0.2.0
configparser>=5.2.0